backend/requirements.txt
absl-py==1.3.0
anyio==3.6.2
astroid==2.12.12
autopep8==2.0.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
chex==0.1.5
click==8.1.3
commonmark==0.9.1
contourpy==1.0.6
cycler==0.11.0
diffusers==0.6.0
dill==0.3.6
dm-tree==0.1.7
fastapi==0.87.0
filelock==3.8.0
flax==0.6.1
fonttools==4.38.0
ftfy==6.1.1
google-api-core==2.10.2
google-auth==2.14.1
google-cloud-core==2.3.2
google-cloud-storage==2.6.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
h11==0.14.0
httptools==0.5.0
huggingface-hub==0.11.0rc1
idna==3.4
importlib-metadata==5.0.0
install==1.3.5
isort==5.10.1
jax==0.3.24
jaxlib==0.3.24
kiwisolver==1.4.4
lazy-object-proxy==1.8.0
matplotlib==3.6.2
mccabe==0.7.0
msgpack==1.0.4
numpy==1.23.4
opt-einsum==3.3.0
optax==0.1.3
packaging==21.3
Pillow==9.3.0
platformdirs==2.5.4
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pydantic==1.10.2
Pygments==2.13.0
pylint==2.15.5
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
rich==12.6.0
rsa==4.9
scipy==1.9.3
six==1.16.0
sniffio==1.3.0
starlette==0.21.0
tokenizers==0.13.2
tomli==2.0.1
tomlkit==0.11.6
toolz==0.12.0
tqdm==4.64.1
transformers==4.24.0
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
uvloop==0.17.0
watchfiles==0.18.1
wcwidth==0.2.5
websockets==10.4
wrapt==1.14.1
zipp==3.10.0
